Make migration tests a separate beast
